Increase the size of program offsets to uint32_t
If a program exceeded the size of a unsigned short the GD would
either crash or get stuck in an infinite loop. Since it's
fairly easy to create such a program the offset size has been
increased to an uint32_t from an unsigned short.

More work is still required to support more strings.

Fixes #1

+++ object.c
@@ -1192,16 +1192,16 @@ free_prog(struct program *progp)
     total_num_prog_blocks--;
 
     /* Free all function names. */
-    for (i=0; i < (int)progp->num_functions; i++)
+    for (i=0; i < progp->num_functions; i++)
 	if (progp->functions[i].name)
 	    free_sstring(progp->functions[i].name);
 
     /* Free all variable names */
-    for (i=0; i < (int)progp->num_variables; i++)
+    for (i=0; i < progp->num_variables; i++)
 	free_sstring(progp->variable_names[i].name);
 
     /* Free all inherited objects */
-    for (i=0; i < (int)progp->num_inherited - 1; i++)
+    for (i=0; i < progp->num_inherited - 1; i++)
     {
 	free_prog(progp->inherit[i].prog);
 	free_sstring(progp->inherit[i].name);
